Skip time zone in test by using Calendar, not Date
A `java.util.Date` has the local time zone embedded in it, so
instantiating it with the seconds since the epoch will always produce a
local time that varies.

Conversely, `java.util.Calendar` works differently; it always gives back
a date (using `Calendar.getTime()`) that matches the numbers passed in,
even if the time zone does not match. We can use this to ignore the time
zone component of the date but get a `Date` instance for the exact
desired date and time.

As a refactoring, extract the mockCondition into a helper method. All of
these use the exact same fixed values, so no need to support overrides
for them just yet.

Discover time zone bugs sooner by running CI in an arbitrary time zone
that might not necessarily be the same as the dev's time zone.

Elsewhere:

- The properties of the test classes can be private, so let them be.
- Moving the fake condition definition out means that we no longer need
to import all the classes needed to make a fake condition.
- The name "mock condition" was not accurate, so rename to "fake
condition".